# Conflicts:
#	Makefile.PL
#	README
#	lib/Libre/Schema/ResultSet/Journalist.pm

+++ lib/Libre/Controller/API/Journalist.pm
@@ -6,10 +6,17 @@ use namespace::autoclean;
 BEGIN { extends 'CatalystX::Eta::Controller::REST' }
 
 with "CatalystX::Eta::Controller::AutoBase";
+with "CatalystX::Eta::Controller::AutoResultPUT";
 
 __PACKAGE__->config(
-    result      => "DB::User",
-    result_cond => { verified => "true" },
+    # AutoBase.
+    result      => "DB::Journalist",
+    result_cond => { 'user.verified' => "true" },
+    result_attr => { join => "user" },
+
+    # AutoResultPUT.
+    object_key     => "journalist",
+    result_put_for => "update",
 );
 
 sub root : Chained('/api/logged') : PathPart('') : CaptureArgs(0) { }
@@ -20,14 +27,32 @@ sub object : Chained('base') : PathPart('') : CaptureArgs(1) {
     my ($self, $c, $user_id) = @_;
 
     my $user = $c->stash->{collection}->find($user_id);
-    if (!$user || !$user->is_journalist()) {
-        $c->detach("/error_404");
-    }
+    $c->detach("/error_404") unless ref $user;
 
+    $c->stash->{is_me} = int($c->user->id == $user->id);
     $c->stash->{journalist} = $user;
 }
 
-sub result : Chained('object') : PathPart('') : Args(0) : ActionClass('REST') { }
+sub result : Chained('object') : PathPart('') : Args(0) : ActionClass('REST') {
+    my ($self, $c) = @_;
+
+    $c->detach("/api/forbidden") unless $c->stash->{is_me};
+}
+
+sub result_GET {
+    my ($self, $c) = @_;
+
+    return $self->status_ok(
+        $c,
+        entity => {
+            ( map { $_ => $c->stash->{journalist}->$_ } qw/address_state address_city address_zipcode address_street address_residence_number cellphone_number/ ),
+
+            ( map { $_ => $c->stash->{journalist}->user->$_ } qw/id email created_at name surname/ ),
+        }
+    );
+}
+
+sub result_PUT { }
 
 __PACKAGE__->meta->make_immutable;
